Subtitle printer and clear functions
Added the functions for subtitle printer and clear, also added a check for how long there has been a silence to possibly understand when a topic changes. Deleting the line is probably not what I want to do at the end but added it for testing purposes. Probably would like to add a parameter for this to create options such like ['Create new paragraph', 'Delete old contents'] etc.